+++ flatlaf-core/src/test/java/com/formdev/flatlaf/ui/TestFlatStyleableValue.java
@@ -625,7 +625,7 @@ public class TestFlatStyleableValue
 		FlatScrollPaneUI ui = (FlatScrollPaneUI) c.getUI();
 
 		// border
-		flatBorder( c, ui );
+		flatScrollPaneBorder( c, ui );
 
 		testBoolean( c, ui, "showButtons", true );
 	}
@@ -968,15 +968,19 @@ public class TestFlatStyleableValue
 		flatBorder( c, ui );
 
 		testInteger( c, ui, "arc", 123 );
-
 		testBoolean( c, ui, "roundRect", true );
+	}
+
+	private void flatScrollPaneBorder( JComponent c, StyleableUI ui ) {
+		flatBorder( c, ui );
+
+		testInteger( c, ui, "arc", 123 );
 	}
 
 	private void flatTextBorder( JComponent c, StyleableUI ui ) {
 		flatBorder( c, ui );
 
 		testInteger( c, ui, "arc", 123 );
-
 		testBoolean( c, ui, "roundRect", true );
 	}
 
@@ -1038,6 +1042,17 @@ public class TestFlatStyleableValue
 		flatBorder( border );
 
 		testValue( border, "arc", 6 );
+		testValue( border, "roundRect", true );
+	}
+
+	@Test
+	void flatScrollPaneBorder() {
+		FlatScrollPaneBorder border = new FlatScrollPaneBorder();
+
+		// FlatScrollPaneBorder extends FlatBorder
+		flatBorder( border );
+
+		testValue( border, "arc", 6 );
 	}
 
 	@Test
@@ -1048,6 +1063,7 @@ public class TestFlatStyleableValue
 		flatBorder( border );
 
 		testValue( border, "arc", 6 );
+		testValue( border, "roundRect", true );
 	}
 
 	@Test